+++ src/Rules/CheckDispatchArgumentTypesCompatibleWithClassConstructorRule.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace NunoMaduro\Larastan\Rules;
 
 use Illuminate\Foundation\Bus\Dispatchable;
+use Illuminate\Foundation\Events\Dispatchable as EventDispatchable;
 use PhpParser\BuilderFactory;
 use PhpParser\Node;
 use PhpParser\Node\Expr\StaticCall;
@@ -17,7 +18,7 @@ use PHPStan\Rules\RuleErrorBuilder;
 use PHPStan\Type\BooleanType;
 
 /** @implements Rule<StaticCall> */
-class CheckJobDispatchArgumentTypesCompatibleWithClassConstructorRule implements Rule
+class CheckDispatchArgumentTypesCompatibleWithClassConstructorRule implements Rule
 {
     /** @var ReflectionProvider */
     private $reflectionProvider;
@@ -25,12 +26,20 @@ class CheckJobDispatchArgumentTypesCompatibleWithClassConstructorRule implements
     /** @var FunctionCallParametersCheck */
     private $check;
 
+    /** @var string */
+    private $dispatchableClass;
+
+    /**
+     * @param  string  $dispatchableClass
+     */
     public function __construct(
         ReflectionProvider $reflectionProvider,
-        FunctionCallParametersCheck $check
+        FunctionCallParametersCheck $check,
+        string $dispatchableClass
     ) {
         $this->reflectionProvider = $reflectionProvider;
         $this->check = $check;
+        $this->dispatchableClass = $dispatchableClass;
     }
 
     /**
@@ -52,16 +61,7 @@ class CheckJobDispatchArgumentTypesCompatibleWithClassConstructorRule implements
 
         $methodName = $node->name->name;
 
-        if (! in_array($methodName, [
-            'dispatch',
-            'dispatchIf',
-            'dispatchUnless',
-            'dispatchSync',
-            'dispatchNow',
-            'dispatchAfterResponse',
-        ],
-            true)
-        ) {
+        if (! in_array($methodName, $this->getAvailableMethods(), true)) {
             return [];
         }
 
@@ -71,9 +71,11 @@ class CheckJobDispatchArgumentTypesCompatibleWithClassConstructorRule implements
 
         $jobClassReflection = $this->reflectionProvider->getClass($scope->resolveName($node->class));
 
-        if (! $jobClassReflection->hasTraitUse(Dispatchable::class)) {
+        if (! $jobClassReflection->hasTraitUse($this->dispatchableClass)) {
             return [];
         }
+
+        $jobOrEvent = $this->dispatchableClass === Dispatchable::class ? 'job' : 'event';
 
         if (! $jobClassReflection->hasConstructor()) {
             $requiredArgCount = 0;
@@ -85,7 +87,7 @@ class CheckJobDispatchArgumentTypesCompatibleWithClassConstructorRule implements
             if (count($node->getArgs()) > $requiredArgCount) {
                 return [
                     RuleErrorBuilder::message(sprintf(
-                        'Job class %s does not have a constructor and must be dispatched without any parameters.',
+                        ucfirst($jobOrEvent).' class %s does not have a constructor and must be dispatched without any parameters.',
                         $jobClassReflection->getDisplayName()
                     ))->build(),
                 ];
@@ -127,20 +129,45 @@ class CheckJobDispatchArgumentTypesCompatibleWithClassConstructorRule implements
             $constructorReflection->getDeclaringClass()->isBuiltin(),
             $node,
             [
-                'Job class '.$classDisplayName.' constructor invoked with %d parameter in '.$classDisplayName.'::'.$methodName.'(), %d required.',
-                'Job class '.$classDisplayName.' constructor invoked with %d parameters in '.$classDisplayName.'::'.$methodName.'(), %d required.',
-                'Job class '.$classDisplayName.' constructor invoked with %d parameter in '.$classDisplayName.'::'.$methodName.'(), at least %d required.',
-                'Job class '.$classDisplayName.' constructor invoked with %d parameters in '.$classDisplayName.'::'.$methodName.'(), at least %d required.',
-                'Job class '.$classDisplayName.' constructor invoked with %d parameter in '.$classDisplayName.'::'.$methodName.'(), %d-%d required.',
-                'Job class '.$classDisplayName.' constructor invoked with %d parameters in '.$classDisplayName.'::'.$methodName.'(), %d-%d required.',
-                'Parameter %s of job class '.$classDisplayName.' constructor expects %s in '.$classDisplayName.'::'.$methodName.'(), %s given.',
+                ucfirst($jobOrEvent).' class '.$classDisplayName.' constructor invoked with %d parameter in '.$classDisplayName.'::'.$methodName.'(), %d required.',
+                ucfirst($jobOrEvent).' class '.$classDisplayName.' constructor invoked with %d parameters in '.$classDisplayName.'::'.$methodName.'(), %d required.',
+                ucfirst($jobOrEvent).' class '.$classDisplayName.' constructor invoked with %d parameter in '.$classDisplayName.'::'.$methodName.'(), at least %d required.',
+                ucfirst($jobOrEvent).' class '.$classDisplayName.' constructor invoked with %d parameters in '.$classDisplayName.'::'.$methodName.'(), at least %d required.',
+                ucfirst($jobOrEvent).' class '.$classDisplayName.' constructor invoked with %d parameter in '.$classDisplayName.'::'.$methodName.'(), %d-%d required.',
+                ucfirst($jobOrEvent).' class '.$classDisplayName.' constructor invoked with %d parameters in '.$classDisplayName.'::'.$methodName.'(), %d-%d required.',
+                'Parameter %s of '.$jobOrEvent.' class '.$classDisplayName.' constructor expects %s in '.$classDisplayName.'::'.$methodName.'(), %s given.',
                 '', // constructor does not have a return type
-                'Parameter %s of job class '.$classDisplayName.' constructor is passed by reference, so it expects variables only',
-                'Unable to resolve the template type %s in instantiation of job class '.$classDisplayName,
+                'Parameter %s of '.$jobOrEvent.' class '.$classDisplayName.' constructor is passed by reference, so it expects variables only',
+                'Unable to resolve the template type %s in instantiation of '.$jobOrEvent.' class '.$classDisplayName,
                 'Missing parameter $%s in call to '.$classDisplayName.' constructor.',
                 'Unknown parameter $%s in call to '.$classDisplayName.' constructor.',
                 'Return type of call to '.$classDisplayName.' constructor contains unresolvable type.',
             ]
         );
+    }
+
+    /** @return non-empty-string[] */
+    private function getAvailableMethods(): array
+    {
+        if ($this->dispatchableClass === Dispatchable::class) {
+            return [
+                'dispatch',
+                'dispatchIf',
+                'dispatchUnless',
+                'dispatchSync',
+                'dispatchNow',
+                'dispatchAfterResponse',
+            ];
+        }
+
+        if ($this->dispatchableClass === EventDispatchable::class) {
+            return [
+                'dispatch',
+                'dispatchIf',
+                'dispatchUnless',
+            ];
+        }
+
+        return [];
     }
 }